rewrite legacy tests with dash_dcc // gut out additional jest testing files // gut out files that serve no purpose
