--- .builderrc
@@ -1,3 +1,0 @@
----
-archetypes:
-  - dash-components-archetype
