+++ .circleci/config.yml
@@ -186,64 +186,6 @@ jobs:
           PYTHON_VERSION: py27
           PERCY_ENABLE: 0
 
-  test-legacy-39: &test-legacy
-    working_directory: ~/project
-    docker:
-      - image: circleci/python:3.9.2-buster-node-browsers
-        auth:
-          username: dashautomation
-          password: $DASH_PAT_DOCKERHUB
-        environment:
-            PYTHON_VERSION: py39
-            PERCY_PARALLEL_TOTAL: -1
-    parallelism: 2
-    steps:
-      - checkout
-      - run: echo $PYTHON_VERSION > ver.txt
-      - restore_cache:
-          key: dep2-{{ checksum ".circleci/config.yml" }}-{{ checksum "ver.txt" }}-{{ checksum "dev-requirements.txt" }}
-      - attach_workspace:
-          at: ~/project
-      - run:
-          name: 🧪 Run Legacy Integration Tests
-          command: |
-            . venv/bin/activate && rm -rf dash_core_components && ls -la
-            set -eo pipefail
-            cd packages && mv dash-*.tar.gz main.tar.gz && ls -la
-            find . -name "dash_*.gz" | xargs pip install -I --progress-bar off --no-cache-dir
-            pip install --no-cache-dir --progress-bar off main.tar.gz[dev,testing]
-            pip list | grep dash | xargs pip show && cd ..
-            echo $(python -V 2>&1) | grep 3. && python -m unittest tests/test_dash_import.py
-            TESTFILES=$(circleci tests glob "tests/test_integration_*.py" | circleci tests split --split-by=timings)
-            pytest --durations=10 --junitxml=test-reports/junit_legacy.xml ${TESTFILES}
-      - store_artifacts:
-          path: test-reports
-      - store_test_results:
-          path: test-reports
-      - store_artifacts:
-          path: /tmp/dash_artifacts
-
-  test-legacy-36:
-    <<: *test-legacy
-    docker:
-      - image: circleci/python:3.6.13-stretch-node-browsers
-        auth:
-          username: dashautomation
-          password: $DASH_PAT_DOCKERHUB
-        environment:
-          PYTHON_VERSION: py36
-          PERCY_ENABLE: 0
-  test-legacy-27:
-    <<: *test-legacy
-    docker:
-      - image: circleci/python:2.7.18-stretch-node-browsers
-        auth:
-          username: dashautomation
-          password: $DASH_PAT_DOCKERHUB
-        environment:
-          PYTHON_VERSION: py27
-          PERCY_ENABLE: 0
-
 workflows:
   version: 2
   python3.9:
@@ -253,13 +195,9 @@ workflows:
       - test-39:
           requires:
             - build-dash-39
-      - test-legacy-39:
-          requires:
-            - build-dash-39
       - percy-finalize:
           requires:
             - test-39
-            - test-legacy-39
 
   python3.6:
     jobs:
@@ -268,17 +206,11 @@ workflows:
       - test-36:
           requires:
             - build-dash-36
-      - test-legacy-36:
-          requires:
-            - build-dash-36
 
   python2.7:
     jobs:
       - lint-unit-27
       - build-dash-27
       - test-27:
-          requires:
-            - build-dash-27
-      - test-legacy-27:
           requires:
             - build-dash-27
