+++ package.json
@@ -21,8 +21,7 @@
     "private::lint.prettier": "prettier --config .prettierrc src/**/*.js --list-different",
     "prepublishOnly": "rm -rf lib && babel src --out-dir lib --copy-files",
     "start": "webpack-serve ./webpack.serve.config.js --open",
-    "test": "run-s -c lint test:legacy test:intg test:pyimport",
-    "test:legacy": "pytest tests/test_integration*.py",
+    "test": "run-s -c lint test:intg test:pyimport",
     "test:intg": "pytest --nopercyfinalize --headless tests/integration",
     "test:pyimport": "python -m unittest tests/test_dash_import.py",
     "uninstall-local": "pip uninstall dash-core-components -y",
