+++ tests/integration/graph/test_graph_responsive.py
@@ -5,7 +5,7 @@ import dash_html_components as html
 import dash_core_components as dcc
 from dash.dependencies import Input, Output, State
 from dash.exceptions import PreventUpdate
-from utils import wait_for
+from dash.testing import wait
 
 
 @pytest.mark.parametrize("responsive", [True, False, None])
@@ -115,23 +115,10 @@ def test_grrs001_graph(dash_dcc, responsive, autosize, height, width, is_respons
         else height
     )
 
-    wait_for(
-        # 500px card minus (100px header + 20px padding) minus (20px padding on container) -> 360px left
-        lambda: dash_dcc.wait_for_element("#graph svg.main-svg").size.get("height", -1)
-        == initial_height,
-        lambda: "initial graph height {}, expected {}".format(
-            dash_dcc.wait_for_element("#graph svg.main-svg").size.get("height", -1),
-            initial_height,
-        ),
-    )
+    # 500px card minus (100px header + 20px padding) minus (20px padding on container) -> 360px left
+    wait.until(lambda: dash_dcc.wait_for_element("#graph svg.main-svg").size.get("height", -1) == initial_height, 3)
+
     dash_dcc.wait_for_element("#resize").click()
 
-    wait_for(
-        # 500px card minus (200px header + 20px padding) minus (20px padding on container) -> 260px left
-        lambda: dash_dcc.wait_for_element("#graph svg.main-svg").size.get("height", -1)
-        == resize_height,
-        lambda: "resized graph height {}, expected {}".format(
-            dash_dcc.wait_for_element("#graph svg.main-svg").size.get("height", -1),
-            resize_height,
-        ),
-    )
+    # 500px card minus (200px header + 20px padding) minus (20px padding on container) -> 260px left
+    wait.until(lambda: dash_dcc.wait_for_element("#graph svg.main-svg").size.get("height", -1) == resize_height, 3)
