--- tests/setupTests.js
@@ -1,5 +1,0 @@
-// setup file
-import { configure } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
-
-configure({ adapter: new Adapter() });
